#16 Add system events contract
This commit creates an `events` module that aims to hold events definitions in the system
that are shared between components.

To keep things simple, the events are defined in pure kotlin's data classes.
Unfortunately I didn't find the way to represent these events in some interface definition language:
- I tried protocol buffers, but it didn't work since reflection is being used and it caused problems with GraalVM native images
- I considered [Async API](https://www.asyncapi.com/), but its generator didn't support pure kotlin/java version and instead only suitable was `spring boot` that is too specific
- There are more I quickly looked, like [Apache Avro](https://avro.apache.org/), but I put it to consider at later point

The system events are produced as a result of user's interaction with commands APIs (the ones in Open API spec):
- user created
- user updated
- user started to follow another user
- user stopped following another user

The processing and storing the events will be implemented in separate commits.