#16 Implement simple local message bus
In order to be able to process messages/events and keep components decoupled
this commit implements local message bus allowing different modules
to publish and subscribe to specific events types.

This is the solution that can work until any of the module became a separate
application and requiring communication through the network.

Alternatives considered:
1. Using external messaging tool like Kafka or RabbitMQ
* It's not needed now, as all components are in single application
* This would bring additional maintenance complexity and effort to build

2. Using Guava or Green Robot Event Bus library
* Both solutions are based on reflection APIs and didn't work in GraalVM native images