Implement separate user projection to be used for validation
In order to make possible to validate if user commands can be executed
it is useful to implement a separate user projection (that could be shared between modules).

This way, instead of providing validation in each of modules, it can be implemented in single place
and reused in others.

The commit implements new `user-query` module that observes `user created` and `user updated` events
and builds projection of current user object. Then the projection is used in `user-commands` module
for validating the commands such as ensuring that:
- non-existent user cannot be updated
- already existent user cannot be created again
- username needs to be unique