Implement separate projection for user follows
In order to make possible to validate if user follow commands can be executed
it is useful to implement a separate projection.

This way, instead of providing validation in each of modules, it can be implemented in single place
and reused in others.

The commit implements new `user-follow` module that observes `user follow` events
and builds projection of follow relation between users.

Then the projection is used in `user-follow-commands` module for validating the commands such as ensuring that there is no need to create follow events when user wants to follow already followed users (and vice versa).