#18 Add db schema creation task for local development
Since Postgres database used in the project requires creation of
schemas and tables, this commit adds Gradle task that executes
sql scripts from all modules for infrastructure run locally.
(Used for development only)

The tasks assume that sql scripts are always under `sql` directory.

This way the whole infrastructure of the project can be prepared in
2 simple tasks: "runLocalInfrastructure" followed by "configureLocalInfrastructure".